requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
pydantic==2.5.0
pyyaml==6.0.1
pypdf==3.17.1